_ARCHIVE/duplicates/requirements (2).txt

click
rich
numpy
pandas
pyyaml
matplotlib
networkx
pytest
ruff
